# apps/core/contracts/openapi-internal.yaml
...
openapi: 3.1.0
info:
  title: CoreERP Control Plane internal API
  version: 1.0.0
  description: |
    Permukaan **app-ke-Core**: seluruh route dengan prefix `/internal/v1`, dipanggil
    aplikasi bisnis dan sistem luar memakai kredensial layanan, bukan session browser.

    Berkas ini **hasil perakitan** dari sumber tulisan tangan di `contracts/internal/`: satu
    berkas per path dan per komponen, dan satu akar per pembaca. Sunting sumbernya, lalu jalankan
    `python contracts/bundle.py` dari `apps/core`; CI menolak berkas hasil yang tidak sama dengan
    sumbernya. Selain berkas gabungan ini, perakit juga menulis satu spesifikasi per pembaca di
    `contracts/terbit/`: satu per domain integrasi (misalnya `integrasi-finance`, terbit di portal
    `/docs`), ditambah `app` dan `pusat-admin`.

    Kontrak ini sengaja terpisah dari `openapi.json`, yang di-generate Scramble dan hanya
    mencakup surface browser Control Plane (`/api/v1`, session cookie). Generator tidak pernah
    memindai prefix ini. Jangan pindahkan isinya ke output generator: kontrak yang dibuat dari
    kode ikut berubah diam-diam saat kodenya berubah.

    Aturan yang mengikat kontrak ini ada pada **Contract decision gate** di
    `.agents/skills/coreerp-architecture/SKILL.md`. Menambah route pada
    `routes/api.php` tanpa menambahkannya di sini adalah perubahan yang belum selesai.

    Tiga jenis pemanggil, masing-masing dengan kredensialnya sendiri:

...
      `AuthenticateAppService`.
...

...

...
servers:
...
security:
- appId: []
  serviceToken: []
  tenantHeader: []
...
paths:
...
  /fiscal-periods:
    get:
      tags:
      - Directory
      operationId: internal.fiscal-periods.resolve
      summary: Tahun buku dan periode yang memuat satu tanggal
      description: |
        Untuk app yang perhitungannya bergantung pada batas tahun buku, misalnya
        penyusutan aset dengan dasar tahun fiskal. Kalender melekat pada legal entity,
        jadi app cukup menyebut legal entity dan tanggal yang ditanyakan.

        Fakta kalender tetap dimiliki Core. App tidak boleh menyalin bulan awal tahun
        buku ke databasenya sendiri, karena salinan itu akan menyimpang saat tenant
        mengubah kalendernya.
      parameters:
      - name: legal_entity_id
        in: query
        required: true
        schema:
          type: string
          format: ulid
      - name: date
        in: query
        required: true
        description: Tanggal yang dicari periodenya, format `YYYY-MM-DD`.
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Kalender, tahun buku, dan periode yang memuat tanggal tersebut.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: object
                    required:
                    - calendar
                    - year
                    - period
                    properties:
                      calendar:
                        type: object
                        required:
                        - id
                        - code
                        - name
                        properties:
                          id:
                            type: string
                            format: ulid
                          code:
                            type: string
                          name:
                            type: string
                      year:
                        type: object
                        required:
                        - id
                        - name
                        - starts_on
                        - ends_on
                        properties:
                          id:
                            type: string
                            format: ulid
                          name:
                            type: string
                          starts_on:
                            type: string
                            format: date
                          ends_on:
                            type: string
                            format: date
                      period:
                        type: object
                        required:
                        - id
                        - ordinal
                        - name
                        - starts_on
                        - ends_on
                        properties:
                          id:
                            type: string
                            format: ulid
                          ordinal:
                            type: integer
                          name:
                            type: string
                          starts_on:
                            type: string
                            format: date
                          ends_on:
                            type: string
                            format: date
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Legal entity belum punya kalender tahun buku, atau tanggal belum tercakup periodenya.
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /units-of-measure/resolve:
    post:
      tags:
      - Directory
      operationId: internal.units-of-measure.resolve
      summary: Resolusi sekumpulan ID satuan
      description: |
        Dipakai app untuk memvalidasi satuan yang disimpannya masih sah. Bersifat
        semua-atau-tidak: satu ID yang tidak ditemukan atau tidak aktif menggagalkan
        seluruh permintaan dengan 422, bukan mengembalikan sebagian hasil.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - unit_ids
              properties:
                unit_ids:
                  type: array
                  minItems: 1
                  maxItems: 100
                  uniqueItems: true
                  items:
                    type: string
                    format: ulid
      responses:
        '200':
          description: Seluruh satuan ditemukan dan aktif.
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UnitOfMeasure'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    integrationClient:
      type: http
      scheme: bearer
      description: |
        Token klien integrasi berbentuk `<client_id>.<secret>`, diterbitkan admin tenant di
        Identity & access › Klien integrasi dan ditampilkan sekali. Token menentukan tenant;
        header tenant yang dikirim bersamanya diabaikan. Klien yang dicabut, alamat asal di luar
        allowlist, atau scope yang kurang ditolak. Rate limit per klien
        (`coreerp.integration_api_rate_limit`, bawaan 120 per menit).
    appId:
      type: apiKey
      in: header
      name: X-CoreERP-App-Id
      description: Identitas app pemanggil, mis. `human-resources`.
    serviceToken:
      type: apiKey
      in: header
      name: X-CoreERP-Service-Token
      description: |
        Service credential berbentuk `<credential_id>.<secret>`. Token lama tanpa
        `credential_id` masih diterima tetapi memaksa pemindaian bcrypt per credential;
        terbitkan ulang.
    tenantHeader:
      type: apiKey
      in: header
      name: X-CoreERP-Tenant-Id
      description: |
        Tenant yang menjadi konteks request. Credential yang terikat tenant hanya sah
        untuk tenant itu. Nilai ini adalah **satu-satunya** sumber `tenant_id`; body dan
        query tidak pernah dipercaya.
    controlPlaneToken:
      type: http
      scheme: bearer
      description: |
        Token bersama yang dipegang pusat admin dan Core (`COREERP_CONTROL_PLANE_TOKEN`).
        Ia menyatakan **proses mana** yang memanggil, bukan operator mana — jejak audit per
        orang tetap harus datang dari sisi konsol. Kosong di sisi Core berarti pemasangan itu
        menolak setiap perintah pusat admin.
...
  responses:
    Forbidden:
      description: |
        Credential tidak sah untuk app/tenant ini, entitlement tidak aktif, placement
        belum `ready`, atau endpoint terbatas pada app lain. Keempatnya sengaja tidak
        dibedakan agar pemanggil tidak dapat memetakan keadaan tenant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: |
        Batas dilampaui. Kuota di-key per pasangan app dan tenant (default 600 request
        per menit, `coreerp.internal_api_rate_limit`), sehingga satu app yang berisik
        tidak dapat menghabiskan kuota app lain.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthenticated:
      description: Header kredensial atau konteks tenant tidak lengkap.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Field request tidak valid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
    Error:
      type: object
      required:
      - message
      properties:
        message:
          type: string
          description: Pesan untuk operator app. Tidak pernah memuat data tenant lain.
...
    UnitOfMeasure:
      type: object
      required:
      - id
      - code
      - name
      - symbol
      - decimal_places
      properties:
        id:
          type: string
          format: ulid
        code:
          type: string
        name:
          type: string
        symbol:
          type:
          - string
          - 'null'
        decimal_places:
          type: integer
    ValidationError:
      type: object
      required:
      - message
      - errors
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...